Added the ability to zoom into the middle of the map
Added a "Zoom" slider that can adjust the scale factor "ZOOM" (between 1 and 0 where 1 is normal). Changing the slider updates the SphereFrame's "unitFrame" with an adjusted angle range for phi and theta. The "Map" class also uses the "ZOOM" variable to adjust the latitude and longitude line thickness. Lastly, an ellipse is drawn in the center of the map to indicate where you will zoom in.