Add functionality to click on a spot on the map and move that spot to the center
Achieved the functionality of clicking on a spot to zoom in on by taking the 3D coordinates of the heightMap at the location of the mouse and then calculating the theta and phi values and then setting them to the correct range on the sliders that control these values. The rotations in SphereFrame originally went rotZ, rotX, and rotZ for theta, phi, and alpha respectively but this made the calculations of the new rotation difficult since rotX slid up and down away from the center of the map. Changing it to rotY allowed the center of the map (on screen) to be the center that moved up and down with phi and streamlined the calculations.